Cleaned up the comments.
Removed a bug: the dangling hydrogens were removed but themselves had to be removed because of the wrong charges. This was leading to trying to remove them twice. The part now that removes the dangling hydrogens is removed. This way at the end, during the stage "disconnected components shoudl be removed", is what removes the dangling hydrogens.