Fix: Do not crash when one of the suptops is reduced to length 0 while correcting the net_zero.
When a small fragment with a high charge mismatch was being tried, a stategy to tackle the charge was not chosen because every one ended with overlap 0. Meaning the suptop was always reduced. The issue was specifically that we do not allow a suptop of length=0 to be a solution. Once we allow that, this is chosen and the suptop is empty and looses to the bigger suptops.