We previously enabled the use of the same atom names. This fixes the PDB tags (B-factor -1 and 1) for NAMD.
Specifically, we previously used unique atom names for .MOL2 and .PDB.
This means that we could always go back and combine the summary JSON
with any of them to recover which atoms were appearing and disappearing.
Well, that's not the case anymore, and ideally the JSON should be updated too.
Here we simply use indices to recover the original atoms.